abandoning mermaid rewrite effort
The effort to generalize the representation of the data turned out to
be more than ChatGPT could handle, and it's just not worth the effort
to implement this at this time when dot does the job of rendering png
images perfectly well. We can revisit this if/when we decide that the
DAG generation should be overhauled in future.